requirements.txt
fastapi==0.88.0
python-multipart==0.0.5
jupyter==1.0.0
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
pyzbar==0.1.9
requests==2.28.1
uvicorn==0.20.0
